Rename cssStatementBlockString to blockString
