+++ src/rules/declaration-block-single-line-max-declarations/index.js
@@ -1,6 +1,6 @@
 import { isNumber } from "lodash"
 import {
-  cssStatementBlockString,
+  blockString,
   cssStatementStringBeforeBlock,
   isSingleLineString,
   report,
@@ -24,7 +24,7 @@ export default function (quantity) {
 
     root.walkRules(rule => {
 
-      if (!isSingleLineString(cssStatementBlockString(rule))) { return }
+      if (!isSingleLineString(blockString(rule))) { return }
       if (!rule.nodes) { return }
 
       const decls = rule.nodes.filter(node => node.type === "decl")
