+++ src/utils/__tests__/blockString-test.js
@@ -1,14 +1,14 @@
 import test from "tape"
 import postcss from "postcss"
-import cssStatementBlockString from "../cssStatementBlockString"
+import blockString from "../blockString"
 
-test("cssStatementBlockString rules", t => {
+test("blockString rules", t => {
   t.equal(postcssCheck("a { color: pink; }"), "{ color: pink; }")
   t.equal(postcssCheck("a {\n\tcolor: pink;\n\ttop: 0;\n}"), "{\n\tcolor: pink;\n\ttop: 0;\n}")
   t.end()
 })
 
-test("cssStatementBlockString at-rules", t => {
+test("blockString at-rules", t => {
   t.equal(postcssCheck("@media print { a { color: pink; } }"), "{ a { color: pink; } }")
   t.equal(
     postcssCheck("@keyframes foo {\n  0% {\n  top: 0;\n}\n\n  100% {\n  top: 10px;\n}\n}\n"),
@@ -18,5 +18,5 @@ test("cssStatementBlockString at-rules", t => {
 
 function postcssCheck(cssString) {
   const root = postcss.parse(cssString)
-  return cssStatementBlockString(root.first)
+  return blockString(root.first)
 }
