+++ src/utils/blockString.js
