+++ src/utils/index.js
@@ -3,7 +3,7 @@ export { default as blurFunctionArguments } from "./blurFunctionArguments"
 export { default as configurationError } from "./configurationError"
 export { default as functionArgumentsSearch } from "./functionArgumentsSearch"
 export { default as cssNodeContextLookup } from "./cssNodeContextLookup"
-export { default as cssStatementBlockString } from "./cssStatementBlockString"
+export { default as blockString } from "./blockString"
 export { default as cssStatementHasBlock } from "./cssStatementHasBlock"
 export { default as cssStatementHasEmptyBlock } from "./cssStatementHasEmptyBlock"
 export { default as cssStatementStringBeforeBlock } from "./cssStatementStringBeforeBlock"
